Move population operators implementations behind feature gate
